pyproject.toml
[tool.poetry]
name = "crewai-agents-solar-news"
version = "0.1.0"
description = "A CrewAI project for solar news analysis"
authors = ["Stas Masevych"]

[tool.poetry.dependencies]
python = "^3.9.6"
crewai = "^0.5.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 